requirements.txt
streamlit==1.55.0
pillow==12.1.1
python-dotenv==1.2.2
google-genai==1.66.0
numpy==2.4.2
sentence-transformers==5.2.3